Remove camera inversion fix again
This approach needs more work. It fixes off vs on board camera rotation being the same direction (X axis), but it's both in the wrong direction. This would require to swap the assignments of CameraLeft and CameraRight retreived from the ini for keyboard and an inversion of the right stick x axis. This again needs an additional check for the park editor. Therefore this patch is removed and shouldn't be patched in the CWalkCameraComponent.
Maybe there is a similar approach for CSkaterCameraComponent